Adding file and refactor presentation for ui
