+++ app/src/main/java/com/example/cekongkir/external/utils/LogUtils.kt
@@ -1,0 +1,12 @@
+/**
+ * Cek Ongkir Android Apps
+ * Copyright (c) 2021. All rights reserved.
+ * @author Created by Nurholis on 4/7/21 01:07 AM
+ * Last modified 10/7/21 16:04 AM by Nurholis
+ * Licensed under the Apache Licence, Version 2.0
+ * May not copy files in their entirety for either distribution or
+ * rewriting of source code.*/
+package com.example.cekongkir.external.utils
+
+object LogUtils {
+}
