+++ app/src/main/res/layout/main_fragment.xml
@@ -13,7 +13,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.main.ui.fragment.MainFragment">
+    tools:context=".presentation.ui.fragment.fragmentcity.FragmentCity">
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
